calibration improved + angle smoothing fixes
Angle smoothing is now compatible with arrays of different lengths and it will filter out all None values. The calibration will now rotate the image according to the average rotation value. This assumes that the camera does not change position. An automatic check for this might be implemented in the future, but doesn't have priority at this point in time.